Routes fixing and tests for API running with the new configurations. - The server.js file is working the same. The changes will NOT affect other files' execution. All tests are passing. - Only crated the app.js file in order for the tests to work without the listening port declaration because was needed. - The package.json file was changed in order to add a new dependency 'supertest' which is the one responsible to run the tests for the API. - Please don't forget to run 'npm install' once this PR is merged.
